+++ libs/multiStep/navigatorButton.js
