+++ src/components/screens/secondPassphrase/secondPassphrase.js
@@ -2,11 +2,11 @@ import React from 'react';
 import { generatePassphrase } from '../../../utils/passphrase';
 import Fees from '../../../constants/fees';
 import FirstStep from './firstStep';
-import MultiStep from '../../shared/multiStep';
 import SummaryStep from './summaryStep';
 import TransactionResult from '../../shared/transactionResult';
 import routes from '../../../constants/routes';
 import styles from './secondPassphrase.css';
+import MultiStep from '../../../../libs/multiStep';
 
 class SecondPassphrase extends React.Component {
   constructor() {
@@ -42,7 +42,7 @@ class SecondPassphrase extends React.Component {
     const { t, account, secondPassphraseRegistered } = this.props;
     return (
       <div className={styles.wrapper}>
-        <MultiStep showNav={false} finalCallback={this.goToWallet}>
+        <MultiStep finalCallback={this.goToWallet}>
           <FirstStep
             t={t}
             goBack={this.backToPreviousPage}
